Refactor duck interfaces, ASMException, JEIDLoadingPlugin
- Move duck interfaces to ducks pkg
- Keep dummy class for INewChunk as passthrough
- Move ASMException to core pkg
- Add dummy JEIDLoadingPlugin for mod compat
