+++ src/main/java/org/dimdev/jeid/JEIDLoadingPlugin.java
@@ -1,0 +1,8 @@
+package org.dimdev.jeid;
+
+/**
+ * Dummy class for reflection by other mods.
+ * @see org.dimdev.jeid.core.JEIDLoadingPlugin
+ */
+public class JEIDLoadingPlugin {
+}
