+++ src/main/java/org/dimdev/jeid/core/JEIDTransformer.java
@@ -6,7 +6,6 @@ import net.minecraft.potion.PotionEffect;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.registry.RegistryNamespaced;
 import net.minecraftforge.fml.relauncher.FMLLaunchHandler;
-import org.dimdev.jeid.ASMException;
 import org.objectweb.asm.ClassReader;
 import org.objectweb.asm.ClassWriter;
 import org.objectweb.asm.Opcodes;
