+++ src/main/java/org/dimdev/jeid/ducks/INewBlockStateContainer.java
@@ -1,7 +1,10 @@
-package org.dimdev.jeid;
+package org.dimdev.jeid.ducks;
 
 import net.minecraft.world.chunk.NibbleArray;
 
+/**
+ * Duck interface for BlockStateContainer mixin.
+ */
 public interface INewBlockStateContainer {
     void setTemporaryPalette(int[] temporaryPalette);
     int[] getTemporaryPalette();
