+++ src/main/java/org/dimdev/jeid/ducks/INewChunk.java
@@ -1,0 +1,9 @@
+package org.dimdev.jeid.ducks;
+
+/**
+ * Duck interface for Chunk mixins.
+ */
+public interface INewChunk {
+    int[] getIntBiomeArray();
+    void setIntBiomeArray(int[] intBiomeArray);
+}
