+++ src/main/java/org/dimdev/jeid/ducks/IStoredEffectInt.java
@@ -1,5 +1,8 @@
 package org.dimdev.jeid.ducks;
 
+/**
+ * Duck interface for SPacketEntityEffect mixin.
+ */
 public interface IStoredEffectInt {
     int getEffectInt();
 }
