+++ src/main/java/org/dimdev/jeid/mixin/modsupport/compactmachines/MixinChunkUtils.java
@@ -16,8 +16,8 @@ import net.minecraftforge.fml.relauncher.ReflectionHelper;
 import org.dave.compactmachines3.utility.ChunkUtils;
 import org.dave.compactmachines3.world.data.provider.AbstractExtraTileDataProvider;
 import org.dave.compactmachines3.world.data.provider.ExtraTileDataProviderRegistry;
-import org.dimdev.jeid.INewBlockStateContainer;
-import org.dimdev.jeid.INewChunk;
+import org.dimdev.jeid.ducks.INewBlockStateContainer;
+import org.dimdev.jeid.ducks.INewChunk;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.Pseudo;
