+++ src/main/java/org/dimdev/jeid/mixin/modsupport/cubicchunks/MixinIONbtWriter.java
@@ -6,9 +6,9 @@ import net.minecraft.nbt.NBTTagList;
 import net.minecraft.world.chunk.Chunk;
 import net.minecraft.world.chunk.NibbleArray;
 import net.minecraft.world.chunk.storage.ExtendedBlockStorage;
-import org.dimdev.jeid.INewBlockStateContainer;
-import org.dimdev.jeid.INewChunk;
-import org.dimdev.jeid.modsupport.cubicchunks.INewCube;
+import org.dimdev.jeid.ducks.INewBlockStateContainer;
+import org.dimdev.jeid.ducks.INewChunk;
+import org.dimdev.jeid.ducks.modsupport.cubicchunks.INewCube;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.Pseudo;
