+++ src/main/java/org/dimdev/jeid/mixin/modsupport/cubicchunks/MixinWorldEncoder.java
@@ -10,8 +10,8 @@ import io.github.opencubicchunks.cubicchunks.core.world.cube.Cube;
 import net.minecraft.network.PacketBuffer;
 import net.minecraft.world.chunk.Chunk;
 import net.minecraft.world.chunk.storage.ExtendedBlockStorage;
-import org.dimdev.jeid.INewChunk;
-import org.dimdev.jeid.modsupport.cubicchunks.INewCube;
+import org.dimdev.jeid.ducks.INewChunk;
+import org.dimdev.jeid.ducks.modsupport.cubicchunks.INewCube;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.Pseudo;
