+++ src/main/java/org/dimdev/jeid/mixin/modsupport/geographicraft/MixinDimensionManager.java
@@ -4,7 +4,7 @@ import climateControl.DimensionManager;
 import net.minecraft.init.Biomes;
 import net.minecraft.world.biome.Biome;
 import net.minecraft.world.chunk.Chunk;
-import org.dimdev.jeid.INewChunk;
+import org.dimdev.jeid.ducks.INewChunk;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.Pseudo;
