+++ src/main/java/org/dimdev/jeid/mixin/modsupport/journeymap/MixinChunkMD.java
@@ -4,7 +4,7 @@ import journeymap.client.model.ChunkMD;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.biome.Biome;
 import net.minecraft.world.chunk.Chunk;
-import org.dimdev.jeid.INewChunk;
+import org.dimdev.jeid.ducks.INewChunk;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Overwrite;
 import org.spongepowered.asm.mixin.Pseudo;
